PR issues resolved
